requirements.txt
# ================================================
# Escalation Helper - Dependencies
# ================================================
# Install with: pip install -r requirements.txt
# ================================================

# Vector database for storing embeddings
chromadb>=0.5.0

# OpenAI for embeddings and LLM
openai>=1.40.0

# Web interface
streamlit>=1.38.0

# Environment variable management
python-dotenv>=1.0.0

# Text processing
tiktoken>=0.7.0

# Markdown parsing
markdown>=3.5.1

# Cross-encoder reranking for improved search accuracy
sentence-transformers>=2.2.0
